minor refactor + dev notes
