+++ package.json
@@ -33,7 +33,7 @@
     "@types/mocha": "^9.1.1",
     "hardhat": "^2.9.9",
     "lerna": "^5.5.4",
-    "nx": "^14.8.4",
+    "nx": "^15.0.0",
     "prettier": "2.7.1",
     "rimraf": "^3.0.2",
     "ts-node": "^10.9.1"
